\name{basic}
\alias{basic}
\docType{data}
\title{
Dataset containing 10 periodic time series
}
\description{
Example dataset as used in the MoPS vignette.
Data matrix with 10 periodic noisy sine wave time courses (rows) with 41 time points (columns).
}
\usage{data(basic)}
\format{
  The format is:
  num [1:10, 1:41] -0.688 -1.237 0.361 1.004 0.775 ... 

}
\keyword{datasets}
