% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mcs.r
\name{mcs_plot_dotty}
\alias{mcs_plot_dotty}
\title{Creates dotty plots from Monte-Carlo simulation output}
\usage{
mcs_plot_dotty(x, annotate = TRUE, outfile = "dotty.pdf",
  overwrite = FALSE)
}
\arguments{
\item{x}{An object of class 'mcs_eval' created by \code{\link{mcs_eval}}.}

\item{annotate}{Logical switch to enable/disable the display of some statistical
measures in the plots.}

\item{outfile}{A name for the output file. The extension '.pdf' is appended
automatically.}

\item{overwrite}{If \code{TRUE}, an existing output file is silently replaced.}
}
\value{
\code{NULL}.
}
\description{
The function takes an output of \code{\link{mcs_eval}} to create
plots of the marginal distribution(s) of the objective function(s) with
respect to the individual parameters. These plots are sometimes referred to
as 'dotty plots'. They provide a means to visually check parameter
sensitivity and identifyability. The plots are saved in a file (pdf format).
}
\note{
If the number of objective functions and/or independend observation data sets is
  too large the layout may be sub-optimal or plotting may fail at all.
  The function is known to work for 3 functions and 3 observation data sets, at least.
}
\author{
David Kneis \email{david.kneis@uni-potsdam.de}
}

