\name{MoPS-internal}

\keyword{internal}
\alias{createLinearTimeCourses}
\alias{decimpart}
\alias{extend}
\alias{fun.lognormal.psi}
\alias{fun.lognormal}
\alias{makewave}
\alias{pathfunction}


\title{Internal MoPS Functions}
\description{Internal MoPS functions }
\details{These are not to be called by the user}
\author{Philipp Eser, Achim Tresch}

