% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mops.r
\docType{package}
\name{mops-package}
\alias{mops}
\alias{mops-package}
\title{Utilities to optimize parameters of dynamic simulation models}
\description{
Type \code{help(package="mops")} to inspect the package description.
}

