% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.r
\name{rc_anyall}
\alias{col_alls}
\alias{col_anys}
\alias{rc_anyall}
\alias{row_alls}
\alias{row_anys}
\title{Matrix Row/Col Any/All}
\usage{
col_anys(x, na.rm = FALSE)

row_anys(x, na.rm = FALSE)

col_alls(x, na.rm = FALSE)

row_alls(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric (integer or double) or logical matrix.}

\item{na.rm}{Should \code{NA} values be removed?}
}
\description{
Row/Column any/all values. Analogues of e.g. \code{colSums()} and \code{rowSums()}.
}
\details{
If \code{na.rm=FALSE}, then \code{NA}'s are ignored.  If, for example, we 
have a column of all \code{NA} values, \code{col_anys()} will return 
\code{FALSE} and \code{col_all()} will return \code{TRUE}.
}

