\name{MoPS-package}
\alias{MoPS-package}
\alias{MoPS}
\docType{package}
\title{
MoPS - Model-based Periodicity Screening
}
\description{
The MoPS package implements the model-based periodicity screen as used in Eser et al.(Mol Syst Biol, 2014) for the detection and characterization of periodic genes.  
}
\details{
\tabular{ll}{
Package: \tab MoPS\cr
Type: \tab Package\cr
Version: \tab 0.99.0\cr
Date: \tab 2014-04-15\cr
License: \tab GPL-3\cr
}
MoPS provides methods for screening numerical time series data for periodicity.
See the package vignette for a detailed description of the methods and recommended workflows. 
}
\author{
Philipp Eser, Achim Tresch

Maintainer: Philipp Eser <eser@genzentrum.lmu.de>

}
\references{
Periodic mRNA synthesis and degradation co-operate during cell cycle gene expression (Eser et al. Mol Sys Biol, 2014) 
}
\keyword{ package }
