% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.r
\name{rc_minmax}
\alias{col_maxs}
\alias{col_mins}
\alias{rc_minmax}
\alias{row_maxs}
\alias{row_mins}
\title{Matrix Row/Col Min/Maxes}
\usage{
col_mins(x)

row_mins(x)

col_maxs(x)

row_maxs(x)
}
\arguments{
\item{x}{A numeric (integer or double) or logical matrix.}
}
\description{
Row/Column minimum/maximum values. Analogues of e.g. \code{colSums()} and \code{rowSums()}.
The functions handle \code{NA}'s appropriately.
}

