% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.r
\name{rc_sums}
\alias{col_sums}
\alias{rc_sums}
\alias{row_sums}
\title{Matrix Row/Col Sums}
\usage{
col_sums(x, na.rm = FALSE)

row_sums(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric (integer or double) or logical matrix.}

\item{na.rm}{Should \code{NA} values be removed?}
}
\description{
Optimized versions of R's \code{rowSums()} and \code{colSums()}.
}

